Bigger data field and better css

+++ database/migrations/2017_05_08_191916_create_stack_folders_table.php
@@ -20,7 +20,7 @@ class CreateStackFoldersTable extends Migration
             $table->string('path_slug', 4096)->unique();
             $table->string('path_hash')->unique();
 
-            $table->integer('size')->unsinged();
+            $table->bigInteger('size')->unsinged();
 
             $table->integer('parent_id')->unsinged()->nullable();
             $table->timestamps();
